Have physical equality inspect Javascript objects
Currently in wasm, physical equality behaves differently from in
Javascript: Javascript objects are boxed, and non-equality of the
pointers causes the values to be physically not equal. This can be an
issue for some folks who rely on the JS-object-inspecting semantics of
physical equality to implement things such as memoizing.

This restores the fact that physical equality inspects Javascript
objects. Other values are unaffected. This entails to pay an additional
cost of between one and two Wasm runtime type tests in physical
equality. A quick benchmark performing a few million `(==)` in an array
of integers show a slowdown of 20~25 %. Since typical programs should
perform physical equality much less frequently, we expect the overhead
to be in the noise in practice.

That being said, we may want to make this behaviour opt-in using a flag.

Co-authored-by: Jérôme Vouillon <jerome.vouillon@gmail.com>